component-entrypoints
- `radio-button`: добавлена точка входа `index.ts` (patch)
- `range`: добавлена точка входа `index.ts` (patch)
- `rating`: добавлена точка входа `index.ts` (patch)
- `side-page`: добавлена точка входа `index.ts` (patch)
- `spinner`: добавлена точка входа `index.ts` (patch)
- `spinner`: константа `DIAMETERS` переименована в `SPINNER_DIAMETER` (major)
