+++ src/radio-button/index.ts
@@ -1,0 +1,2 @@
+export type { RadioButtonProps } from './types';
+export { RadioButton } from './radio-button';
