+++ src/range/index.ts
@@ -1,0 +1,2 @@
+export type { RangeProps } from './types';
+export { Range } from './range';
