+++ src/rating/index.ts
@@ -1,0 +1,2 @@
+export type { RatingProps } from './types';
+export { Rating } from './rating';
