+++ src/side-page/index.ts
@@ -1,0 +1,3 @@
+export type { SidePageProps } from './types';
+export { SidePage } from './side-page';
+export { SidePageHeader, SidePageBody, SidePageFooter } from './slots';
