+++ src/spinner/index.ts
@@ -1,0 +1,2 @@
+export type { SpinnerSize, SpinnerProps } from './types';
+export { Spinner, SpinnerSVG, SPINNER_DIAMETER } from './spinner';
